requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
persiantools>=3.0.1
sqlite3 
aiohttp==3.9.5
aiosignal==1.3.1
altgraph @ file:///AppleInternal/Library/BuildRoots/226e9c8c-edb1-11ee-8f17-a65dcee5a99e/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.6.0
anyio==3.7.0
appnope==0.1.4
APScheduler==3.6.3
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
black==24.2.0
blessed==1.20.0
blinker==1.7.0
Brotli==1.0.9
cachetools==4.2.2
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
Django==4.2.11
docopt==0.6.2
editor==1.6.6
exceptiongroup==1.1.1
executing==2.0.1
filelock==3.15.3
flake8==7.0.0
Flask==2.2.2
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
fonttools==4.42.0
frozenlist==1.4.1
fsspec==2024.6.0
ftfy==6.2.0
future @ file:///AppleInternal/Library/BuildRoots/226e9c8c-edb1-11ee-8f17-a65dcee5a99e/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
html2image==2.0.4.3
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.28.1
huggingface-hub==0.23.4
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
importlib_resources==6.4.0
iniconfig==2.0.0
inquirer==3.2.5
instaloader==4.13.1
ipykernel==6.29.4
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
litellm==1.40.8
macholib @ file:///AppleInternal/Library/BuildRoots/226e9c8c-edb1-11ee-8f17-a65dcee5a99e/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.4.4
markdown-it-py==3.0.0
markdown2==2.4.10
Markdown2PDF==0.1.4
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
md2pdf==1.0.1
mdurl==0.1.2
mpmath==1.3.0
multidict==4.7.6
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
open-interpreter==0.3.3
openai==1.35.3
packaging==24.1
parso==0.8.4
pathspec==0.12.1
peft==0.7.0
persiantools==5.1.0
pexpect==4.9.0
Pillow==10.0.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
pydyf==0.7.0
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pyphen==0.14.0
pytest==8.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.0.3
readchar==4.1.0
regex==2024.5.15
requests==2.31.0
requests-async==0.5.0
rfc3986==1.5.0
rich==13.7.1
rsa==4.9
runs==1.2.2
safetensors==0.4.3
sanic==19.6.0
Send2Trash==1.8.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.36
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.12.1
Telethon==1.32.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.19.1
tokentrim==0.1.13
toml==0.10.2
tomli==2.2.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
transformers==4.42.3
typing_extensions==4.11.0
tzlocal==5.2
ujson==5.9.0
urllib3==1.23
uvloop==0.19.0
wcwidth==0.2.13
weasyprint==59.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==6.0
Werkzeug==2.1.2
wget==3.2
xmod==1.8.1
yarl==1.9.4
zipp==3.16.2
zopfli==0.2.2
